Refactoring of main and code cleanup
