+++ scan_core/src/transition_system.rs
@@ -166,7 +166,7 @@ pub trait TransitionSystem: Clone + Send + Sync {
                     // disregard state and move on.
                     return Continue((s, f));
                 } else if guarantees.iter().all(|p| p.eval(trace.as_slice())) {
-                    // If all guarantees are satisfied,
+                    // If all guarantees are satisfied, the execution is successful
                     s += 1;
                 } else {
                     // If guarantee is violated, we have found a counter-example!
@@ -200,8 +200,8 @@ pub trait TransitionSystem: Clone + Send + Sync {
                 self.clone().experiment(guarantees, assumes, &mut rng)
             })
             .inspect(|result| {
-                if *result && rand::thread_rng().gen_bool(0.666) {
-                    // If all guarantees are satisfied,
+                if *result {
+                    // If all guarantees are satisfied, the execution is successful
                     let mut s = s.fetch_add(1, Ordering::Relaxed);
                     s += 1;
                     info!("runs: {s} successes");
@@ -225,23 +225,31 @@ pub trait TransitionSystem: Clone + Send + Sync {
 // Carlos E. Budde, Pedro R. D’Argenio, Arnd Hartmanns, Sean Sedwards.
 // International Journal on Software Tools for Technology Transfer (2020) 22:759–780
 // https://doi.org/10.1007/s10009-020-00563-2
+
+/// Computes Okamoto bound for given confidence and precision.
 pub fn okamoto_bound(confidence: f64, precision: f64) -> f64 {
     (2f64 / (1f64 - confidence)).ln() / (2f64 * precision.powf(2f64))
 }
 
+/// Computes adaptive bound for given confidence, precision and (partial) experimental results.
 pub fn adaptive_bound(s: u32, f: u32, confidence: f64, precision: f64) -> f64 {
     let n = s + f;
+    // Avoid division by 0
     let avg = s as f64 / n.max(1) as f64;
     4f64 * okamoto_bound(confidence, precision)
         * (0.25f64 - ((avg - 0.5f64).abs() - (2f64 * precision / 3f64)).powf(2f64))
 }
 
+/// Computes precision for given experimental results and confidence
+/// deriving it from adaptive bound through quadratic equation.
 pub fn derive_precision(s: u32, f: u32, confidence: f64) -> f64 {
     let n = s + f;
     let avg = s as f64 / n as f64;
     let k = 2f64 * (2f64 / (1f64 - confidence)).ln();
+    // Compute quadratic equation coefficients.
     let a = (n as f64) + (4f64 * k / 9f64);
-    let b = -4f64 * k / 3f64 * ((avg - 0.5f64).abs());
+    let b = -4f64 * k * (avg - 0.5f64).abs() / 3f64;
     let c = k * ((avg - 0.5f64).powf(2f64) - 0.25f64);
+    // Take (larger positive) quadratic equation solution.
     (-b + (b.powf(2f64) - 4f64 * a * c).sqrt()) / (2f64 * a)
 }
